refactor(editor): decompose EditorPane into single-responsibility sub-modules
Split EditorPane.tsx (~1517 lines) into ≤300-line sub-modules:

- EditorPane.tsx (256): layout orchestration + re-exports
- useEditorSetup.ts (218): TipTap lifecycle, content loading, autosave
- editorPasteUtils.ts (186): paste handling, sanitization, capacity utils
- InlineAiOverlay.tsx (200): inline AI input/proposal UI
- EntityCompletionPopover.tsx (47): @mention popover component
- entityCompletionUtils.ts (97): entity completion detection + matching
- useEntityCompletion.ts (268): entity completion hook
- SlashCommandMenu.tsx (71): slash panel state + wrapper
- useEditorKeybindings.ts (240): keyboard shortcuts + action handlers
- EditorToolbar.tsx (82): toolbar orchestration
- ToolbarFormatGroup.tsx (88): inline format buttons
- ToolbarBlockGroup.tsx (212): block-level toolbar buttons
- EditorBubbleMenu.tsx (141): bubble menu visibility/placement
- BubbleMenuFormatActions.tsx (183): bubble menu format + link UI
- BubbleMenuAiActions.tsx (103): bubble menu AI skill buttons
- typography.ts (47): typography variable resolution

Pure structural refactoring — zero functional changes.
All tests pass (290 files, 2071 tests), typecheck clean, lint 0 errors.

Closes #1171
